WIP: Using the new interface for parser
